springboot example package renaming
