+++ examples/spring-boot-example/src/test/java/com/github/qzagarese/dockerunit/examples/springboot/ContainerTest.java
@@ -1,0 +1,5 @@
+package com.github.qzagarese.dockerunit.examples.springboot;
+
+public interface ContainerTest {
+
+}
