+++ examples/spring-boot-example/src/test/java/com/github/qzagarese/dockerunit/examples/springboot/descriptors/BaseDescriptor.java
@@ -1,10 +1,10 @@
-package com.github.qzagarese.microunit.examples.springboot.descriptors;
+package com.github.qzagarese.dockerunit.examples.springboot.descriptors;
 
 import com.github.qzagarese.dockerunit.annotation.Image;
 import com.github.qzagarese.dockerunit.annotation.Named;
 import com.github.qzagarese.dockerunit.annotation.PublishPorts;
 import com.github.qzagarese.dockerunit.discovery.consul.annotation.WebHealthCheck;
-import com.github.qzagarese.microunit.examples.springboot.Constants;
+import com.github.qzagarese.dockerunit.examples.springboot.Constants;
 
 @Named(Constants.SERVICE_NAME)
 @Image(Constants.IMAGE_NAME)
