+++ examples/spring-boot-example/src/test/java/com/github/qzagarese/dockerunit/examples/springboot/descriptors/TestEnvDescriptor.java
@@ -1,14 +1,14 @@
-package com.github.qzagarese.microunit.examples.springboot.descriptors;
+package com.github.qzagarese.dockerunit.examples.springboot.descriptors;
 
-import static com.github.qzagarese.microunit.examples.springboot.Constants.BAR_VALUE_FROM_DESCRIPTOR;
-import static com.github.qzagarese.microunit.examples.springboot.Constants.FOO_VALUE_FROM_DESCRIPTOR;
+import static com.github.qzagarese.dockerunit.examples.springboot.Constants.BAR_VALUE_FROM_DESCRIPTOR;
+import static com.github.qzagarese.dockerunit.examples.springboot.Constants.FOO_VALUE_FROM_DESCRIPTOR;
 
 import com.github.qzagarese.dockerunit.annotation.Env;
 import com.github.qzagarese.dockerunit.annotation.Image;
 import com.github.qzagarese.dockerunit.annotation.Named;
 import com.github.qzagarese.dockerunit.annotation.PortBinding;
 import com.github.qzagarese.dockerunit.discovery.consul.annotation.WebHealthCheck;
-import com.github.qzagarese.microunit.examples.springboot.Constants;
+import com.github.qzagarese.dockerunit.examples.springboot.Constants;
 
 @Named(Constants.SERVICE_NAME)
 @Image(Constants.IMAGE_NAME)
