--- examples/spring-boot-example/src/test/java/com/github/qzagarese/microunit/examples/springboot/ContainerTest.java
@@ -1,5 +1,0 @@
-package com.github.qzagarese.microunit.examples.springboot;
-
-public interface ContainerTest {
-
-}
